K19DKPA-63 adaptation to new ddl

+++ backend/src/main/java/edu/netcracker/backend/model/TicketClass.java
@@ -15,10 +15,10 @@ public class TicketClass {
 
     @PrimaryKey("class_id")
     @EqualsAndHashCode.Include
-    private Long classId;
+    private Integer classId;
 
     @Attribute("trip_id")
-    private Long tripId;
+    private Integer tripId;
 
     @Attribute("ticket_price")
     private Integer ticketPrice;
